requirements.txt
facebook-business==25.0.0
python-dotenv==1.0.1
python-telegram-bot
anthropic
gdown
